Pedantic changes
Might sound better with just one "fewer".

"Examples of default exports may be a single StorePicker React Component or an array of data." is a really weird sentence. 

Changed `var` to `const` for `dogs`, but I guess that's something of an opinion.

"a dozen or two" is a little weird (sounds like "12 or 2", not "12 or 24")

I made a handful of English stylistic changes. I tried not to change the tone or your style, but I found some of the wording a little too awkward so I reworded them. Hope you don't mind. =)