Babel transpiles to CommonJS by default
Babel is not a bundler. It's a transpiler. The bundler is the one that leverages the `node_modules/` directory.

This is significant and Babel 6 actually [changed/broke behavior](http://stackoverflow.com/q/33505992/971592) on how it transpiled export statements so this is an important distinction to make. Because these modules are consumable via CommonJS you can consume them as named exports OR default exports. But that's not how their code is being written. They're simply doing `module.exports = _` or `module.exports = jQuery`.